Resolve MakeGenericType ILLink warning in DependencyInjection
This adds RequiresUnreferencedCode to any DI API that could take a separate Service and Implementation type. The reason is because if the Implementation type is generic and has a DynamicallyAccessedMembers annotation on its generic types, and the Service (interface) type doesn't, the trimmer won't preserve the correct members on the inner type. This leads to runtime failures.

Adding the RequiresUnreferencedCode attribute allows us to suppress the MakeGenericType warning we are getting in DependencyInjection, but it forces all places that add DI services to ensure their types annotations match, and then suppress the warning.